landing_page - added load more button to product-tabs component

+++ backend/model/ProductReview.js
@@ -10,7 +10,7 @@ const ProductReview = db.define(
             autoIncrement: true
         },
         rating: {
-            type: Sequelize.INTEGER(1),
+            type: Sequelize.SMALLINT,
             allowNull: false
         },
         productId: {
